Change from `standard` to `eslint`
This commit configures ESLint to enforce the JavaScript Standard Style, which is used in this project. ESLint will be used instead of JavaScript Standard Style’s `standard` cli tool.

Currently, the new ESLint “flat config” format is not supported by `eslint-config-standard`; as a workaround, the compatible rules from `eslint-config-standard` are directly included in the configuration file instead.