Revamped verification process
This commit revamps the process for verifying a Discord account. In addition, it migrates the codebase to openid-client@6.x.

Rather than storing information in encrypted and authenticated cookies, the IdentiBot web interface now stores a session ID cookie, which contains a base64url-encoded, pseudorandom 256-bit value. The session ID cookie points to relevant information in the MongoDB database that would formerly have been stored in cookies. The cookie expires after 12 hours, and the associated record remains in the database for 12.5 hours after creation, at which time it should be pruned using the MongoDB TTL feature. When verification is completed, the record is removed from the database and the cookie from the browser.

Furthermore, the async/promise handling during the verification flow has been improved in this commit.