Reimplement auto name changing
This commit reimplements automatic nickname changing for servers that have opted in to this feature. The feature was temporarily removed in c03e11ff.

A small change has also been made to the code that pauses for a random amount of time before changes are applied to newly verified users across servers (part of the rate limit avoidance mechanism).