Automatically assign roles when running /verify
This commit makes IdentiBot automatically assign the appropriate roles to a user when they run the /verify command in a server based on accounts that they have already linked. This allows users to get the appropriate roles in a server newly configured for use with IdentiBot without needing to leave and rejoin.